python-django/requirements.txt
asgiref==3.11.0
Django==6.0.1
djangorestframework==3.16.1
drf-spectacular==0.29.0
drf-spectacular-sidecar==2026.1.1
onyx-database==0.0.4
sqlparse==0.5.5
